Fixed direct message crash

+++ commands/approve.js
@@ -3,6 +3,11 @@ var applications = require('../util/applications.js');
 
 module.exports = {
     execute: function(args, message) {
+        if(!message.guild) {
+            message.author.send("Sorry, but this command doesn't work in direct messages!");
+            return;
+        }
+
         if(!message.guild.member(message.author.id).hasPermission('ADMINISTRATOR')) {
             message.author.send("Only admins of a server may use the !approve command! And you are no admin, sorry :/");
             message.delete();
